--- plugin/oov/mecab_oov/Cargo.toml
@@ -1,17 +1,0 @@
-[package]
-name = "mecab_oov"
-version = "0.1.0"
-authors = ["Works Applications <sudachi@worksap.co.jp>"]
-edition = "2018"
-license = "Apache-2.0"
-
-[lib]
-crate-type = ["cdylib"]
-
-[dependencies]
-serde = { version = "1.0", features = ["derive"] }
-serde_json = "1.0"
-sudachi = { path = "../../../sudachi" }
-
-[dev-dependencies]
-tempfile = "3"
