+++ sudachi-cli/Cargo.toml
@@ -14,8 +14,6 @@ sudachi = { path = "../sudachi" }
 structopt = "0.3"
 cfg-if = "1.0.0"
 
-mecab_oov = { path = "../plugin/oov/mecab_oov" }
-simple_oov = { path = "../plugin/oov/simple_oov" }
 join_numeric = { path = "../plugin/path_rewrite/join_numeric" }
 join_katakana_oov = { path = "../plugin/path_rewrite/join_katakana_oov" }
