+++ sudachi/src/stateless_tokeniser.rs
@@ -35,7 +35,7 @@ pub trait DictionaryAccess {
     fn grammar(&self) -> &Grammar<'_>;
     fn lexicon(&self) -> &LexiconSet<'_>;
     fn input_text_plugins(&self) -> &[Box<dyn InputTextPlugin + Sync + Send>];
-    fn oov_provider_plugins(&self) -> &[Box<dyn OovProviderPlugin + Sync>];
+    fn oov_provider_plugins(&self) -> &[Box<dyn OovProviderPlugin + Sync + Send>];
     fn path_rewrite_plugins(&self) -> &[Box<dyn PathRewritePlugin + Sync>];
 }
